requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-rest-swagger==2.2.0
django-safedelete==1.0.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.1
Faker==6.6.2
firebase-admin==4.5.2
google-api-core==1.26.2
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
msgpack==1.0.2
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
proto-plus==1.18.1
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
